requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
python-decouple>=3.8
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn==1.5.1
python-dotenv>=1.0.0
google-genai>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0
xgboost>=2.0.0
joblib>=1.3.0
tensorflow>=2.15.0